pyproject.toml
[project]
name = "avans-ai-security-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic",
    "ipykernel>=7.2.0",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
]
